X, Y, Z are now working (#16)
* Added `nmax` value to tests.

* Maximum spherical harmonic degree is returned as well now.

* Maximal spherical harmonic degree is handled now.

* Fixed typo.

* Switching to float64 inputs for `igrf` package.

* Fixed automatic tests creation.

* shval3 initial revision.

* WIP: some calculations are started.

* Fixed error with incorrect `date`.

* Tests are now working for X, Y, Z

* Some clean up

* Making linter happy.

+++ igrf/igrf_edge_test.go
@@ -7,10 +7,10 @@ import (
 
 func TestIGRFEdgeCases(t *testing.T) {
 	type args struct {
-		lat  float32
-		lon  float32
-		alt  float32
-		date float32
+		lat  float64
+		lon  float64
+		alt  float64
+		date float64
 	}
 	tests := []struct {
 		name    string
@@ -54,6 +54,12 @@ func TestIGRFEdgeCases(t *testing.T) {
 			want:    IGRFresults{},
 			wantErr: true,
 		},
+		// {
+		// 	name:    "Testing",
+		// 	args:    args{lat: -59.9, lon: -39.9, alt: -0.5, date: 1915.5},
+		// 	want:    IGRFresults{},
+		// 	wantErr: false,
+		// },
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
